requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
psycopg2-binary==2.9.6
requests==2.31.0
soupsieve==2.4.1
urllib3==2.0.3
apache-airflow==2.2.2
